email: do not close the input attachment buffer in Conversion#parseBodyFields
Currently, Conversion#parseBodyFields close the inputstream associated to
the temporary binary attachment stream after reading the textContent and the snipet.
This will prevent to copy to the destination database folder for inline attachments.
Just let the attachment to be closed for LegacyConversions#addOneAttachment method.

Requires: http://review.cyanogenmod.org/#/c/98128

Change-Id: Ibe41e1896d6867cae0ee080f1556d7fa9e88966d
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>