Don't re-show search bar on query click.
If we click query the search bar must be visible already. The actual
problem with this, however, is that this is triggered via the default
implementation of onKeyUp in the query view, which means that it's
triggered when pressing enter on the keyboard, but _after_ the submit
IME action is delivered.

Change-Id: I4b3d7ec3c0040a3ac7cf8460f8059f5074f9149d